map/autoware_lanelet2_divider/requirements.txt
pyyaml==6.0.1
lark==0.11.3
pytest==7.1.3
utm==0.7.0
mgrs==1.4.6
gdal==3.4
tqdm==4.66.2
lanelet2==1.2.1